feat(hogql): move macro and asterisk expansion to resolver
