+++ posthog/hogql/hogql.py
@@ -18,15 +18,14 @@ def translate_hogql(query: str, context: HogQLContext, dialect: Literal["hogql",
     try:
         # Create a fake query that selects from "events" to have fields to select from.
         context.database = context.database or create_hogql_database(context.team_id)
-        select_query_type = ast.SelectQueryType(tables={"events": ast.TableType(table=context.database.events)})
         node = parse_expr(query, no_placeholders=True)
-        select_query = ast.SelectQuery(
-            select=[node], select_from=ast.JoinExpr(table=ast.Field(chain=["events"])), type=select_query_type
-        )
-        prepared_ast: ast.SelectQuery = cast(
+        select_query = ast.SelectQuery(select=[node], select_from=ast.JoinExpr(table=ast.Field(chain=["events"])))
+        prepared_select_query: ast.SelectQuery = cast(
             ast.SelectQuery,
             prepare_ast_for_printing(select_query, context=context, dialect=dialect, stack=[select_query]),
         )
-        return print_prepared_ast(prepared_ast.select[0], context=context, dialect=dialect, stack=[select_query])
+        return print_prepared_ast(
+            prepared_select_query.select[0], context=context, dialect=dialect, stack=[prepared_select_query]
+        )
     except (NotImplementedException, SyntaxException):
         raise
