--- posthog/hogql/transforms/__init__.py
@@ -1,2 +1,0 @@
-from .asterisk import expand_asterisks
-from .lazy_tables import resolve_lazy_tables
