--- posthog/hogql/transforms/test/test_asterisk.py
@@ -1,199 +1,0 @@
-from posthog.hogql import ast
-from posthog.hogql.database import create_hogql_database
-from posthog.hogql.parser import parse_select
-from posthog.hogql.resolver import ResolverException, resolve_types
-from posthog.hogql.transforms import expand_asterisks
-from posthog.test.base import BaseTest
-
-
-class TestAsteriskExpander(BaseTest):
-    def setUp(self):
-        self.database = create_hogql_database(self.team.pk)
-
-    def test_asterisk_expander_table(self):
-        node = parse_select("select * from events")
-        node = resolve_types(node, self.database)
-        expand_asterisks(node)
-        events_table_type = ast.TableType(table=self.database.events)
-        self.assertEqual(
-            node.select,
-            [
-                ast.Field(chain=["uuid"], type=ast.FieldType(name="uuid", table_type=events_table_type)),
-                ast.Field(chain=["event"], type=ast.FieldType(name="event", table_type=events_table_type)),
-                ast.Field(chain=["properties"], type=ast.FieldType(name="properties", table_type=events_table_type)),
-                ast.Field(chain=["timestamp"], type=ast.FieldType(name="timestamp", table_type=events_table_type)),
-                ast.Field(chain=["distinct_id"], type=ast.FieldType(name="distinct_id", table_type=events_table_type)),
-                ast.Field(
-                    chain=["elements_chain"], type=ast.FieldType(name="elements_chain", table_type=events_table_type)
-                ),
-                ast.Field(chain=["created_at"], type=ast.FieldType(name="created_at", table_type=events_table_type)),
-            ],
-        )
-
-    def test_asterisk_expander_table_alias(self):
-        node = parse_select("select * from events e")
-        node = resolve_types(node, self.database)
-        expand_asterisks(node)
-        events_table_type = ast.TableType(table=self.database.events)
-        events_table_alias_type = ast.TableAliasType(table_type=events_table_type, alias="e")
-        self.assertEqual(
-            node.select,
-            [
-                ast.Field(chain=["uuid"], type=ast.FieldType(name="uuid", table_type=events_table_alias_type)),
-                ast.Field(chain=["event"], type=ast.FieldType(name="event", table_type=events_table_alias_type)),
-                ast.Field(
-                    chain=["properties"], type=ast.FieldType(name="properties", table_type=events_table_alias_type)
-                ),
-                ast.Field(
-                    chain=["timestamp"], type=ast.FieldType(name="timestamp", table_type=events_table_alias_type)
-                ),
-                ast.Field(
-                    chain=["distinct_id"], type=ast.FieldType(name="distinct_id", table_type=events_table_alias_type)
-                ),
-                ast.Field(
-                    chain=["elements_chain"],
-                    type=ast.FieldType(name="elements_chain", table_type=events_table_alias_type),
-                ),
-                ast.Field(
-                    chain=["created_at"], type=ast.FieldType(name="created_at", table_type=events_table_alias_type)
-                ),
-            ],
-        )
-
-    def test_asterisk_expander_subquery(self):
-        node = parse_select("select * from (select 1 as a, 2 as b)")
-        node = resolve_types(node, self.database)
-        expand_asterisks(node)
-        select_subquery_type = ast.SelectQueryType(
-            aliases={
-                "a": ast.FieldAliasType(alias="a", type=ast.ConstantType(data_type="int")),
-                "b": ast.FieldAliasType(alias="b", type=ast.ConstantType(data_type="int")),
-            },
-            columns={
-                "a": ast.FieldAliasType(alias="a", type=ast.ConstantType(data_type="int")),
-                "b": ast.FieldAliasType(alias="b", type=ast.ConstantType(data_type="int")),
-            },
-            tables={},
-            anonymous_tables=[],
-        )
-        self.assertEqual(
-            node.select,
-            [
-                ast.Field(chain=["a"], type=ast.FieldType(name="a", table_type=select_subquery_type)),
-                ast.Field(chain=["b"], type=ast.FieldType(name="b", table_type=select_subquery_type)),
-            ],
-        )
-
-    def test_asterisk_expander_subquery_alias(self):
-        node = parse_select("select x.* from (select 1 as a, 2 as b) x")
-        node = resolve_types(node, self.database)
-        expand_asterisks(node)
-        select_subquery_type = ast.SelectQueryAliasType(
-            alias="x",
-            select_query_type=ast.SelectQueryType(
-                aliases={
-                    "a": ast.FieldAliasType(alias="a", type=ast.ConstantType(data_type="int")),
-                    "b": ast.FieldAliasType(alias="b", type=ast.ConstantType(data_type="int")),
-                },
-                columns={
-                    "a": ast.FieldAliasType(alias="a", type=ast.ConstantType(data_type="int")),
-                    "b": ast.FieldAliasType(alias="b", type=ast.ConstantType(data_type="int")),
-                },
-                tables={},
-                anonymous_tables=[],
-            ),
-        )
-        self.assertEqual(
-            node.select,
-            [
-                ast.Field(chain=["a"], type=ast.FieldType(name="a", table_type=select_subquery_type)),
-                ast.Field(chain=["b"], type=ast.FieldType(name="b", table_type=select_subquery_type)),
-            ],
-        )
-
-    def test_asterisk_expander_from_subquery_table(self):
-        node = parse_select("select * from (select * from events)")
-        node = resolve_types(node, self.database)
-        expand_asterisks(node)
-
-        events_table_type = ast.TableType(table=self.database.events)
-        inner_select_type = ast.SelectQueryType(
-            tables={"events": events_table_type},
-            anonymous_tables=[],
-            aliases={},
-            columns={
-                "uuid": ast.FieldType(name="uuid", table_type=events_table_type),
-                "event": ast.FieldType(name="event", table_type=events_table_type),
-                "properties": ast.FieldType(name="properties", table_type=events_table_type),
-                "timestamp": ast.FieldType(name="timestamp", table_type=events_table_type),
-                "distinct_id": ast.FieldType(name="distinct_id", table_type=events_table_type),
-                "elements_chain": ast.FieldType(name="elements_chain", table_type=events_table_type),
-                "created_at": ast.FieldType(name="created_at", table_type=events_table_type),
-            },
-        )
-
-        self.assertEqual(
-            node.select,
-            [
-                ast.Field(chain=["uuid"], type=ast.FieldType(name="uuid", table_type=inner_select_type)),
-                ast.Field(chain=["event"], type=ast.FieldType(name="event", table_type=inner_select_type)),
-                ast.Field(chain=["properties"], type=ast.FieldType(name="properties", table_type=inner_select_type)),
-                ast.Field(chain=["timestamp"], type=ast.FieldType(name="timestamp", table_type=inner_select_type)),
-                ast.Field(chain=["distinct_id"], type=ast.FieldType(name="distinct_id", table_type=inner_select_type)),
-                ast.Field(
-                    chain=["elements_chain"],
-                    type=ast.FieldType(name="elements_chain", table_type=inner_select_type),
-                ),
-                ast.Field(chain=["created_at"], type=ast.FieldType(name="created_at", table_type=inner_select_type)),
-            ],
-        )
-
-    def test_asterisk_expander_multiple_table_error(self):
-        node = parse_select("select * from (select 1 as a, 2 as b) x left join (select 1 as a, 2 as b) y on x.a = y.a")
-        with self.assertRaises(ResolverException) as e:
-            node = resolve_types(node, self.database)
-        self.assertEqual(
-            str(e.exception), "Cannot use '*' without table name when there are multiple tables in the query"
-        )
-
-    def test_asterisk_expander_select_union(self):
-        node = parse_select("select * from (select * from events union all select * from events)")
-        node = resolve_types(node, self.database)
-        expand_asterisks(node)
-
-        events_table_type = ast.TableType(table=self.database.events)
-        inner_select_type = ast.SelectUnionQueryType(
-            types=[
-                ast.SelectQueryType(
-                    tables={"events": events_table_type},
-                    anonymous_tables=[],
-                    aliases={},
-                    columns={
-                        "uuid": ast.FieldType(name="uuid", table_type=events_table_type),
-                        "event": ast.FieldType(name="event", table_type=events_table_type),
-                        "properties": ast.FieldType(name="properties", table_type=events_table_type),
-                        "timestamp": ast.FieldType(name="timestamp", table_type=events_table_type),
-                        "distinct_id": ast.FieldType(name="distinct_id", table_type=events_table_type),
-                        "elements_chain": ast.FieldType(name="elements_chain", table_type=events_table_type),
-                        "created_at": ast.FieldType(name="created_at", table_type=events_table_type),
-                    },
-                )
-            ]
-            * 2
-        )
-
-        self.assertEqual(
-            node.select,
-            [
-                ast.Field(chain=["uuid"], type=ast.FieldType(name="uuid", table_type=inner_select_type)),
-                ast.Field(chain=["event"], type=ast.FieldType(name="event", table_type=inner_select_type)),
-                ast.Field(chain=["properties"], type=ast.FieldType(name="properties", table_type=inner_select_type)),
-                ast.Field(chain=["timestamp"], type=ast.FieldType(name="timestamp", table_type=inner_select_type)),
-                ast.Field(chain=["distinct_id"], type=ast.FieldType(name="distinct_id", table_type=inner_select_type)),
-                ast.Field(
-                    chain=["elements_chain"],
-                    type=ast.FieldType(name="elements_chain", table_type=inner_select_type),
-                ),
-                ast.Field(chain=["created_at"], type=ast.FieldType(name="created_at", table_type=inner_select_type)),
-            ],
-        )
